fixup! Refactor cause adding new class MailboxAnnotationKey
