+++ protocols/imap/src/test/java/org/apache/james/imap/decode/ImapRequestLineReaderTest.java
@@ -1,0 +1,13 @@
+package org.apache.james.imap.decode;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class ImapRequestLineReaderTest {
+
+    @Test
+    public void testNextNonSpaceChar() throws Exception {
+
+    }
+}
